style: move Footer
